+++ settings.lua
@@ -1,0 +1,10 @@
+data:extend({
+  {
+    type = "bool-setting",
+    name = "tfc_state",
+    setting_type = "runtime-global",
+    default_value = true,
+    localised_name = {"gui-mod-info.toggle"},
+    localised_description  = ""
+  }
+})