+++ CHANGELOG.md
@@ -1,9 +1,19 @@
 # Changelog
 
+## 2019-01-19
+
+### [v0.3.0][v0.3.0]
+
+- Added: localization
+- Added: setting for state of mod
+- [Fixed](https://mods.factorio.com/mod/timesaver-for-crafting/discussion/5c42cd347f3064000b047d3d): error when craft an item in cheat mode
+- Updated: embedded module «[Event listener](https://gitlab.com/ZwerOxotnik/event-listener)» to 0.1.1
+
 ## 2019-01-02
 
-### [v1.0.0][v1.0.0]
+### [v0.1.0][v0.1.0]
 
 - first release for 0.16
 
-[v1.0.0]: https://mods.factorio.com/download/timesaver-for-crafting/5c2cdd91f64c7c000b0d2e84
+[v0.3.0]: https://mods.factorio.com/mod/timesaver-for-crafting/downloads
+[v0.1.0]: https://mods.factorio.com/download/timesaver-for-crafting/5c2cdd91f64c7c000b0d2e84