+++ docs/ru/CHANGELOG.md
@@ -1,9 +1,19 @@
 # Список изменений
 
+## 2019-01-19
+
+### [v0.3.0][v0.3.0]
+
+- Добавлено: локализация
+- Добавлено: настройки состояния мода
+- [Исправлено](https://mods.factorio.com/mod/timesaver-for-crafting/discussion/5c42cd347f3064000b047d3d): ошибка при созданни предмета в читерском режиме
+- Обновлен: встроенный модуль «[Прослушиватель событий](https://gitlab.com/ZwerOxotnik/event-listener)» до 0.1.1
+
 ## 2019-01-02
 
-### [v1.0.0][v1.0.0]
+### [v0.1.0][v0.1.0]
 
 - первый выпуск для 0.16
 
-[v1.0.0]: https://mods.factorio.com/download/timesaver-for-crafting/5c2cdd91f64c7c000b0d2e84
+[v0.3.0]: https://mods.factorio.com/mod/timesaver-for-crafting/downloads
+[v0.1.0]: https://mods.factorio.com/download/timesaver-for-crafting/5c2cdd91f64c7c000b0d2e84